+++ NEWS.md
@@ -12,10 +12,12 @@ Enhancements:
 Bug fixes:
 
 * plot_surv_animated now uses gganimate::transition_manual() internally, ruling out interpolation errors
+* There are no slight differences in the outlines of the areas when using plot_surv_area with discrete=TRUE and discrete=FALSE anymore
 
 New Features:
 
 * Added plot_surv_matrix function
 * Added support for bootstrap standard error and confidence interval calculation in the curve_cont function
-* Added support for bootstrap confidence intervals in the plot_surv_at_t, plot_surv_lines, plot_surv_animated functions
+* Added support for bootstrap confidence intervals in the plot_surv_at_t, plot_surv_lines and plot_surv_animated functions
+* Added support for plotting Kaplan-Meier curves as reference in plot_surv_area, plot_surv_animated, plot_surv_lines
 * Vignette now includes plot_surv_matrix function